requirements.txt
tweepy
python-telegram-bot